% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_response}
\alias{check_response}
\title{Check response from PLOS, including status codes, server error messages, mime-type, etc.}
\usage{
check_response(x)
}
\description{
Check response from PLOS, including status codes, server error messages, mime-type, etc.
}
\keyword{internal}
