% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{densitylist}
\alias{densitylist}
\title{The density web service provides access to records showing the density
    of occurrence records from the GBIF Network by one-degree cell.}
\usage{
densitylist(...)
}
\description{
This function is defunct.
}
\keyword{internal}

