% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{gbifmap_list}
\alias{gbifmap_list}
\title{Make a simple map to visualize GBIF point data.}
\usage{
gbifmap_list(...)
}
\description{
This function is defunct.
}
\seealso{
gbifmap
}
\keyword{internal}

