% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{get_credentials}
\alias{get_credentials}
\title{Get Github credentials from use in console}
\usage{
get_credentials(...)
}
\description{
This function is defunct.  See the package gistr for similar functionality.
}
\keyword{internal}

