% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{insertnones}
\alias{insertnones}
\title{Function to insert "none" character strings where NULL values found to
faciliate combining results}
\usage{
insertnones(x)
}
\description{
Function to insert "none" character strings where NULL values found to
faciliate combining results
}
\keyword{internal}

