% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalnamekey.R
\name{journalnamekey}
\alias{journalnamekey}
\title{Get short keys for journals to use in searching specific journals.}
\usage{
journalnamekey()
}
\value{
(character) journal name keys
}
\description{
Get short keys for journals to use in searching specific journals.
}
\examples{
\dontrun{
journalnamekey()
}
}
