% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{occurrencecount}
\alias{occurrencecount}
\title{Counts taxon concept records matching a range of filters.}
\usage{
occurrencecount(...)
}
\description{
This function is defunct.
}
\seealso{
occ_count
}
\keyword{internal}

