% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{occurrenceget}
\alias{occurrenceget}
\title{Get individual records for a given occurrence record.}
\usage{
occurrenceget(...)
}
\description{
This function is defunct.
}
\seealso{
occ_get
}
\keyword{internal}

