% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{occurrencelist_all}
\alias{occurrencelist_all}
\title{Occurrencelist_all carries out an occurrencelist query for a single name and
all its name variants according to GBIF's name matching.}
\usage{
occurrencelist_all(...)
}
\description{
This function is defunct.
}
\seealso{
occ_search
}
\keyword{internal}

