% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{occurrencelist_many}
\alias{occurrencelist_many}
\title{occurrencelist_many is the same as occurrencelist, but takes in a vector
of species names.}
\usage{
occurrencelist_many(...)
}
\description{
This function is defunct.
}
\seealso{
occ_search
}
\keyword{internal}

