% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plosword.R
\name{plosword}
\alias{plosword}
\title{Search results on a keyword over all fields in PLoS Journals.}
\usage{
plosword(terms, vis = FALSE, ...)
}
\arguments{
\item{terms}{search terms (character)}

\item{vis}{visualize results in bar plot or not (TRUE or FALSE)}

\item{...}{Optional additional curl options passed to
\code{\link[crul]{HttpClient}}}
}
\value{
Number of search results (vis = FALSE), or number of search in a
table and a histogram of results (vis = TRUE).
}
\description{
Search results on a keyword over all fields in PLoS Journals.
}
\examples{
\dontrun{
plosword('Helianthus')
plosword(list('monkey','replication','design','sunflower','whale'),
   vis = TRUE)
}
}
