% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{providers}
\alias{providers}
\title{Get data providers and their unique keys.}
\usage{
providers(...)
}
\description{
This function is defunct.
}
\seealso{
networks organizations datasets
}
\keyword{internal}

