% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{resources}
\alias{resources}
\title{Get data resources and their unique keys.}
\usage{
resources(...)
}
\description{
This function is defunct.
}
\seealso{
networks organizations datasets
}
\keyword{internal}

