% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rplos-package.R
\docType{package}
\name{rplos}
\alias{rplos}
\alias{rplos-package}
\title{Connect with PLoS API data}
\description{
rplos provides a wrapper to the PLoS Search API and the Article-Level
metrics API.  More information about each function can be found in
its help documentation.
}
\details{
rplos functions

Most rplos functions make web calls (using the httr package), and either
parse xml (using the XML package) or json (using the jsonlite package)
results.

PLoS API keys

You used to need an API key to use this package - no longer as of 2015-01-13

rplos tutorial at rOpenSci website here:
http://ropensci.org/tutorials/rplos-tutorial/

Tutorials

See the rOpenSci website for a tutorial:
http://ropensci.org/tutorials/rplos_tutorial.html
}

