% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{stylegeojson}
\alias{stylegeojson}
\title{Style a data.frame prior to converting to geojson.}
\usage{
stylegeojson(...)
}
\description{
This function is defunct.  See the package spocc for similar functionality.
}
\keyword{internal}

