% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{taxoncount}
\alias{taxoncount}
\title{Search by taxon to retrieve number of records in GBIF.}
\usage{
taxoncount(...)
}
\description{
This function is defunct.
}
\seealso{
occ_count
}
\keyword{internal}

