% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{taxonget}
\alias{taxonget}
\title{Get taxonomic information on a specific taxon or taxa in GBIF by their taxon
  	concept keys.}
\usage{
taxonget(...)
}
\description{
This function is defunct.
}
\seealso{
name_usage
}
\keyword{internal}

