% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defunct.R
\name{togeojson}
\alias{togeojson}
\title{Convert spatial data files to GeoJSON from various formats.}
\usage{
togeojson(...)
}
\description{
This function is defunct.  See the package togeojson for similar functionality.
}
\keyword{internal}

