\name{concat_todf}
\alias{concat_todf}
\title{Concatenate author names, if present, used in other functions.}
\usage{
  concat_todf(x)
}
\arguments{
  \item{x}{a single list element with PLoS API returned
  nested elements}
}
\value{
  data.frame of results, with authors concatenated to
  single vector.
}
\description{
  Concatenate author names, if present, used in other
  functions.
}

