\name{formatcrossref}
\alias{formatcrossref}
\title{Convert crossref XML into a bibentry object}
\usage{
  formatcrossref(a)
}
\arguments{
  \item{a}{crossref XML output}
}
\value{
  a bibentry format of the output
}
\description{
  Convert crossref XML into a bibentry object
}
\details{
  internal helper function
}

