% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fulltext.R
\name{full_text_urls}
\alias{full_text_urls}
\title{Create urls for full text articles in PLOS journals.}
\usage{
full_text_urls(doi)
}
\arguments{
\item{doi}{One or more doi's}
}
\value{
One or more urls, same length as input vector of dois
}
\description{
Create urls for full text articles in PLOS journals.
}
\examples{
\dontrun{
full_text_urls(doi='10.1371/journal.pone.0086169')
full_text_urls(doi='10.1371/journal.pbio.1001845')
full_text_urls(doi=c('10.1371/journal.pone.0086169','10.1371/journal.pbio.1001845'))
dois <- searchplos(q = "*:*", fq='doc_type:full', limit=20)$data$id
full_text_urls(dois)
}
}

