% Generated by roxygen2 (4.0.0): do not edit by hand
\name{insertnones}
\alias{insertnones}
\title{Function to insert "none" character strings where NULL values found to
faciliate combining results}
\usage{
insertnones(x)
}
\description{
Function to insert "none" character strings where NULL values found to
faciliate combining results
}
\keyword{internal}

