% Generated by roxygen2 (4.0.1): do not edit by hand
\name{journalnamekey}
\alias{journalnamekey}
\title{Get short keys for journals to use in searching specific journals.}
\usage{
journalnamekey()
}
\value{
Journals name keys.
}
\description{
Get short keys for journals to use in searching specific journals.
}
\examples{
\dontrun{
journalnamekey()
}
}

