\name{journalnamekey}
\alias{journalnamekey}
\title{Get short keys for journals to use in searching specific journals.}
\usage{
  journalnamekey()
}
\value{
  Journals name keys.
}
\description{
  Get short keys for journals to use in searching specific
  journals.
}
\examples{
\dontrun{
journalnamekey()
}
}

