% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_throughtime.R
\name{plot_throughtime}
\alias{plot_throughtime}
\title{Plot results through time for serach results from PLoS Journals.}
\usage{
plot_throughtime(terms, limit = NA, key = getOption("PlosApiKey",
  stop("need an API key for PLoS Journals")), ...)
}
\arguments{
\item{terms}{search terms (character)}

\item{limit}{number of results to return (integer)}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}

\item{...}{optional additional curl options (debugging tools mostly)}
}
\value{
Number of search results (vis = FALSE), or number of search in a table
   and a histogram of results (vis = TRUE).
}
\description{
Plot results through time for serach results from PLoS Journals.
}
\examples{
\dontrun{
plot_throughtime(terms='phylogeny', limit=300)
plot_throughtime(list('drosophila','monkey'), 100)
plot_throughtime(list('drosophila','flower','dolphin','cell','cloud'), 100)
}
}

